Moving all py files in emod-api to down_case.py style from DownCase.py (Fix #33)
