+++ emod_api/demographics/demographics_generator.py
@@ -9,8 +9,8 @@ from emod_api.dtk_tools.demographics.DemographicsGeneratorConcern import (
     DemographicsGeneratorConcern,
     DemographicsGeneratorConcernChain,
 )
-from emod_api.demographics.Node import Node, nodeid_from_lat_lon
-from emod_api.demographics.Demographics import Demographics
+from emod_api.demographics.node import Node, nodeid_from_lat_lon
+from emod_api.demographics.demographics import Demographics
 from emod_api.dtk_tools.support.General import init_logging
 
 logger = init_logging("DemographicsGenerator")
