+++ emod_api/demographics/fertility_distribution.py
@@ -2,7 +2,7 @@ from typing import List, Dict
 
 import emod_api.demographics.demographic_exceptions as demog_ex
 
-from emod_api.demographics.Updateable import Updateable
+from emod_api.demographics.updateable import Updateable
 from emod_api.utils import check_dimensionality
