+++ emod_api/demographics/mortality_distribution.py
@@ -1,7 +1,7 @@
 from typing import List, Dict, Union
 
 import emod_api.demographics.demographic_exceptions as demog_ex
-from emod_api.demographics.Updateable import Updateable
+from emod_api.demographics.updateable import Updateable
 from emod_api.utils import check_dimensionality
