+++ emod_api/demographics/mortality_distribution_old.py
@@ -1,6 +1,6 @@
 from typing import List, Any
 
-from emod_api.demographics.Updateable import Updateable
+from emod_api.demographics.updateable import Updateable
 
 
 class MortalityDistributionOld(Updateable):
